fix:  Switch to wsdl generated WsSMC
